+++ x-pack/plugin/ql/src/main/java/org/elasticsearch/xpack/ql/expression/Foldables.java
@@ -1,0 +1,18 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+package org.elasticsearch.xpack.ql.expression;
+
+import org.elasticsearch.xpack.ql.QlIllegalArgumentException;
+
+public abstract class Foldables {
+
+    public static Object valueOf(Expression e) {
+        if (e.foldable()) {
+            return e.fold();
+        }
+        throw new QlIllegalArgumentException("Cannot determine value for {}", e);
+    }
+}
