+++ x-pack/plugin/sql/src/main/java/org/elasticsearch/xpack/sql/planner/QueryTranslator.java
@@ -10,6 +10,7 @@ import org.elasticsearch.geometry.Point;
 import org.elasticsearch.search.sort.SortOrder;
 import org.elasticsearch.xpack.ql.expression.Expression;
 import org.elasticsearch.xpack.ql.expression.FieldAttribute;
+import org.elasticsearch.xpack.ql.expression.Foldables;
 import org.elasticsearch.xpack.ql.expression.Literal;
 import org.elasticsearch.xpack.ql.expression.NamedExpression;
 import org.elasticsearch.xpack.ql.expression.function.Function;
@@ -81,6 +82,7 @@ import org.elasticsearch.xpack.sql.querydsl.agg.TopHitsAgg;
 import org.elasticsearch.xpack.sql.type.SqlDataTypeConverter;
 import org.elasticsearch.xpack.sql.util.Check;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -88,8 +90,7 @@ import java.util.Set;
 
 import static java.util.Collections.singletonList;
 import static org.elasticsearch.xpack.ql.expression.Expressions.id;
-import static org.elasticsearch.xpack.sql.expression.Foldables.doubleValuesOf;
-import static org.elasticsearch.xpack.sql.expression.Foldables.valueOf;
+import static org.elasticsearch.xpack.ql.expression.Foldables.valueOf;
 
 final class QueryTranslator {
 
@@ -636,7 +637,7 @@ final class QueryTranslator {
 
         @Override
         protected LeafAgg toAgg(String id, Percentiles p) {
-            return new PercentilesAgg(id, field(p), doubleValuesOf(p.percents()));
+            return new PercentilesAgg(id, field(p), foldAndConvertToDoubles(p.percents()));
         }
     }
 
@@ -644,7 +645,7 @@ final class QueryTranslator {
 
         @Override
         protected LeafAgg toAgg(String id, PercentileRanks p) {
-            return new PercentileRanksAgg(id, field(p), doubleValuesOf(p.values()));
+            return new PercentileRanksAgg(id, field(p), foldAndConvertToDoubles(p.values()));
         }
     }
 
@@ -696,5 +697,13 @@ final class QueryTranslator {
         }
 
         protected abstract LeafAgg toAgg(String id, C f);
+    }
+    
+    private static List<Double> foldAndConvertToDoubles(List<Expression> list) {
+        List<Double> values = new ArrayList<>(list.size());
+        for (Expression e : list) {
+            values.add((Double) SqlDataTypeConverter.convert(Foldables.valueOf(e), DataTypes.DOUBLE));
+        }
+        return values;
     }
 }
